m68k: fix remaining -mfastcall gaps from round 5 review
bios/arch/m68k/aciavecs.S: mfpint() call had no -mfastcall branch and
still packed num as a tightly-packed WORD.

bios/arch/m68k/processor.S: set_cache() had no -mfastcall branch; the
incoming enable arg (in d0) needed stashing in d1 before _cache_exists
and the moveq below clobber d0.

bios/arch/m68k/68040_pmmu.S: balloc_stram() call had no -mfastcall
branch (size/top need to move to d0/d1, and its UBYTE* return comes
back in a0, not d0), and BOOL top's plain-ABI slot needed widening now
that BOOL is a 32-bit int.

bios/arch/m68k/natfeat.S: _xhdi_vec's call into xhdi_handler() had no
-mfastcall branch for its pointer argument.

bdos/console.[ch], bdos/fs.h, bdos/mem.h, bdos/fsdir.c, bdos/fsfat.c,
bdos/fshand.c, bdos/fsio.c, bdos/fsmain.c, bdos/fsopnclo.c, bdos/umem.c:
several GEMDOS entry points (xconout, xauxout, xprtout, xrawio,
xsetdrv, xforce, xdup, xclose, xopen, xgetfree, xlseek, xread, xwrite,
xgsdtof, xrename, xgetdir, xsetblk, xmxalloc) were declared with `int`
scalar parameters but dispatched through bdosmain.c's shape table via
a `short`-typed function-pointer union member. Under -mfastcall only
the low word of such an argument register is guaranteed meaningful;
the real (int-parameter) function reads the full register, so garbage
upper bits could corrupt the value (e.g. xrawio's Crawio(0xff)
sentinel). Their true wire contract is a single WORD, so retype them
to WORD to match the union member they're actually called through.

Kconfig.machine: restrict CONF_WITH_MFASTCALL to ARCH_M68K_CLASSIC.
ColdFire's vdi_tblit.S is a generated file (see "make coldfire-sources"
in the Makefile), machine-translated by a proprietary tool this project
can't run in CI; its _normal_blit has not been audited or re-generated
for -mfastcall's a0/a1 register-argument convention, so disable the
combination instead of leaving it silently broken.

Verified: atari192, atari256 (both -mfastcall) and firebee (ColdFire,
plain ABI) build clean; atari192 boots to the desktop under Hatari.

+++ bdos/fshand.c
@@ -29,7 +29,7 @@
  * Error returns:
  *     EIHNDL
  */
-long xforce(int std, int h)
+long xforce(WORD std, WORD h)
 {
     return ixforce(std,h,run);
 }
@@ -90,7 +90,7 @@ long ixforce(int std, int h, PD *p)
  *     ENHNDL
  *
  */
-long xdup(int h)
+long xdup(WORD h)
 {
     int i;
     long fh;
